adding spring-data-jpa to reduce boilerplate JPA criteria code

+++ core/src/main/java/com/sopovs/moradanen/fan/service/IDaoService.java
@@ -38,7 +38,7 @@ public interface IDaoService {
 
     List<PlayerInGame> lastGamesForPlayer(Long playerId, int size, int startFrom);
 
-    int countGames();
+    long countGames();
 
     int countGamesForTeam(Long teamId);
